+++ tools/common/BitCodeMode.cs
@@ -1,0 +1,8 @@
+namespace Xamarin.Bundler {
+	public enum BitCodeMode {
+		None = 0,
+		ASMOnly = 1,
+		LLVMOnly = 2,
+		MarkerOnly = 3,
+	}
+}
