+++ tools/dotnet-linker/dotnet-linker.csproj
@@ -182,6 +182,9 @@
     <Compile Include="..\..\external\Xamarin.MacDev\Xamarin.MacDev\PListObject.cs">
       <Link>external\Xamarin.MacDev\Xamarin.MacDev\PListObject.cs</Link>
     </Compile>
+    <Compile Include="..\common\BitCodeMode.cs">
+      <Link>external\tools\common\BitCodeMode.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="..\mtouch\Errors.resx">
